backend/requirement.txt
fastapi
uvicorn
pymysql
pydantic_settings
sqlalchemy
asyncpg
python-multipart
passlib
pyjwt